requirements.txt
openai
feedparser
beautifulsoup4
python-dotenv
requests
